fix(forkm): behavior in linked worktrees
Correctly set and export GIT_DIR, GIT_COMMON_DIR, GIT_WORK_TREE.

In linked worktrees, .git is a file pointing to a subdirectory of the
main worktree's .git directory (which is GIT_COMMON_DIR), and the
correct value of GIT_DIR is that subdirectory. And, as follows from the
above, GIT_WORK_TREE is not 'GIT_DIR/..'.