fix(unpatch): merge commit selection
Change selection rule to properly reflect what I intended in the first
place.

The problem with the old one was that if the user does `forkm update
fork master` then 'master' does not point to the merge commit, since we
merge with --no-ff - 'master's position is unchanged by the merge. So
all merge commits into 'fork' from before this `forkm update` are not
reachable from 'master', so they're also selected, even though they're
not children of 'master'.

The change to the tests really should have been made in the previous
commit, and they only worked there because the old selection rule was
weird as described above.